requirements.txt
asgiref==3.8.1
Django==5.1.3
django-appconf==1.0.6
django-colorfield==0.11.0
django-filter==24.3
django-jazzmin==3.0.1
django3-admin-select2==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
